requirements.txt
alembic==1.13.3
blinker==1.8.2
click==8.1.7
envparse==0.2.0
flask==3.0.3
Flask-Migrate==4.0.7
flask-sqlalchemy==3.1.1
greenlet==3.1.1
importlib-metadata==8.5.0
itsdangerous==2.2.0
jinja2==3.1.4
Mako==1.3.6
MarkupSafe==3.0.2
marshmallow==3.23.0
packaging==24.1
psycopg2-binary==2.9.10
python-dotenv==1.0.1
SQLAlchemy==2.0.36
SQLAlchemy-Utils==0.41.2
typing-extensions==4.12.2
werkzeug==3.0.6
zipp==3.20.2
